fix: refactor fsd atchitecture
