+++ src/shared/api/customer/index.ts
@@ -1,18 +1,12 @@
 import { supabase } from '@shared/api/supabase-client';
 import type { Customer } from '@shared/interfaces';
 
-import { mapCustomer, mapAddress } from './mapper';
+import { mapAddress, mapCustomer } from './mapper';
 
 export const getCustomer = async (): Promise<Customer | null> => {
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-  if (!user?.id) return null;
-
   const { data: customer } = await supabase
     .from('customers')
     .select('*')
-    .eq('user_id', user.id)
     .maybeSingle()
     .throwOnError();
 
@@ -23,7 +17,6 @@ export const getCustomer = async (): Promise<Customer | null> => {
   const { data: addresses } = await supabase
     .from('addresses')
     .select('*')
-    .eq('user_id', user.id)
     .throwOnError();
 
   const shippingAddresses = mapAddress(
